# Conflicts:
#	Best Recipes.xcodeproj/project.xcworkspace/xcuserdata/alexander.xcuserdatad/UserInterfaceState.xcuserstate

+++ Best Recipes/Flows/FavoriteFlow/Favorites Screen/RecipeModel.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct RecipeModel {
     let id: Int
-    let score: Int
+    let score: Double
     let title: String
     let image: String
-    let cuisines: [String]  
+    let cuisines: [String] 
 }
